requirements.txt
alembic==1.9.0
amqp==5.1.1
apt-wrapper==1.18
bcrypt==3.2.0
billiard==3.6.4.0
blinker==1.4
celery==5.2.7
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==3.4.8
distro==1.7.0
fasteners==0.14.1
Flask==2.2.2
Flask-Login==0.6.2
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.2
future==0.18.2
greenlet==2.0.1
httplib2==0.20.2
idna==2.10
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
kombu==5.2.4
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.1
monotonic==1.6
more-itertools==8.10.0
mysqlclient==1.4.6
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
prompt-toolkit==3.0.36
protobuf==3.12.4
ptyprocess==0.7.0
pycparser==2.21
PyJWT==2.3.0
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-dateutil==2.8.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
schedule==1.1.0
SecretStorage==3.3.1
six==1.16.0
SQLAlchemy==1.4.45
urllib3==1.26.5
vine==5.0.0
wadllib==1.3.6
wcwidth==0.2.5
Werkzeug==2.2.2
xdg==5.0.0
zipp==1.0.0